fixed bug in SAL layer, miscalculating LBA

+++ libc/include/string.h
@@ -11,7 +11,7 @@ int memcmp(const void*, const void*, size_t);
 void* memcpy(void* __restrict, const void* __restrict, size_t);
 void* memmove(void*, const void*, size_t);
 void* memset(void*, int, size_t);
-size_t strlen(const char*);
+int strlen(const char*);
 char* strncat(char* __restrict, const char* __restrict, size_t);
 
 #ifdef __cplusplus
