+++ libc/string/strlen.c
@@ -1,6 +1,6 @@
 #include "../include/string.h"
 
-size_t strlen(const char* str){
+int strlen(const char* str){
     size_t len = 0;
     while (str[len])
         ++len;
